+++ packages/video_player/video_player_avfoundation/lib/src/avfoundation_video_player.dart
@@ -188,10 +188,9 @@ class AVFoundationVideoPlayer extends VideoPlayerPlatform {
         ),
         'completed' => VideoEvent(eventType: VideoEventType.completed),
         'bufferingUpdate' => VideoEvent(
-          buffered:
-              (map['values'] as List<dynamic>)
-                  .map<DurationRange>(_toDurationRange)
-                  .toList(),
+          buffered: (map['values'] as List<dynamic>)
+              .map<DurationRange>(_toDurationRange)
+              .toList(),
           eventType: VideoEventType.bufferingUpdate,
         ),
         'bufferingStart' => VideoEvent(
@@ -227,10 +226,9 @@ class AVFoundationVideoPlayer extends VideoPlayerPlatform {
         textureId: textureId,
       ),
       VideoPlayerPlatformViewState() => _buildPlatformView(playerId),
-      null =>
-        throw Exception(
-          'Could not find corresponding view type for playerId: $playerId',
-        ),
+      null => throw Exception(
+        'Could not find corresponding view type for playerId: $playerId',
+      ),
     };
   }