+++ script/tool/lib/src/gradle_check_command.dart
@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'dart:math' as math;
+
 import 'package:collection/collection.dart';
 import 'package:file/file.dart';
 import 'package:meta/meta.dart';
@@ -24,6 +26,8 @@ class GradleCheckCommand extends PackageLoopingCommand {
     super.packagesDir, {
     super.gitDir,
   });
+
+  static const int _minimumJavaVersion = 17;
 
   @override
   final String name = 'gradle-check';
@@ -126,6 +130,12 @@ class GradleCheckCommand extends PackageLoopingCommand {
       succeeded = false;
     }
     if (!_validateCompatibilityVersions(lines)) {
+      succeeded = false;
+    }
+    if (!_validateKotlinJvmCompatibility(lines)) {
+      succeeded = false;
+    }
+    if (!_validateJavaKotlinCompileOptionsAlignment(lines)) {
       succeeded = false;
     }
     if (!_validateGradleDrivenLintConfig(package, lines)) {
@@ -357,30 +367,32 @@ build.gradle "namespace" must match the "package" attribute in AndroidManifest.x
     final bool hasLanguageVersion = gradleLines.any((String line) =>
         line.contains('languageVersion') && !_isCommented(line));
     final bool hasCompabilityVersions = gradleLines.any((String line) =>
-            line.contains('sourceCompatibility') && !_isCommented(line)) &&
+            line.contains('sourceCompatibility = JavaVersion.VERSION_') &&
+            !_isCommented(line)) &&
         // Newer toolchains default targetCompatibility to the same value as
         // sourceCompatibility, but older toolchains require it to be set
         // explicitly. The exact version cutoff (and of which piece of the
         // toolchain; likely AGP) is unknown; for context see
         // https://github.com/flutter/flutter/issues/125482
         gradleLines.any((String line) =>
-            line.contains('targetCompatibility') && !_isCommented(line));
+            line.contains('targetCompatibility = JavaVersion.VERSION_') &&
+            !_isCommented(line));
     if (!hasLanguageVersion && !hasCompabilityVersions) {
-      const String errorMessage = '''
-build.gradle must set an explicit Java compatibility version.
+      const String javaErrorMessage = '''
+build.gradle(.kts) must set an explicit Java compatibility version.
 
 This can be done either via "sourceCompatibility"/"targetCompatibility":
     android {
         compileOptions {
-            sourceCompatibility = JavaVersion.VERSION_11
-            targetCompatibility = JavaVersion.VERSION_11
+            sourceCompatibility = JavaVersion.VERSION_$_minimumJavaVersion
+            targetCompatibility = JavaVersion.VERSION_$_minimumJavaVersion
         }
     }
 
 or "toolchain":
     java {
         toolchain {
-            languageVersion = JavaLanguageVersion.of(11)
+            languageVersion = JavaLanguageVersion.of($_minimumJavaVersion)
         }
     }
 
@@ -389,9 +401,87 @@ https://docs.gradle.org/current/userguide/java_plugin.html#toolchain_and_compati
 for more details.''';
 
       printError(
-          '$indentation${errorMessage.split('\n').join('\n$indentation')}');
+          '$indentation${javaErrorMessage.split('\n').join('\n$indentation')}');
       return false;
     }
+
+    return true;
+  }
+
+  bool _validateKotlinJvmCompatibility(List<String> gradleLines) {
+    bool isKotlinOptions(String line) =>
+        line.contains('kotlinOptions') && !_isCommented(line);
+    final bool hasKotlinOptions = gradleLines.any(isKotlinOptions);
+    final bool kotlinOptionsUsesJavaVersion = gradleLines.any((String line) =>
+        line.contains('jvmTarget = JavaVersion.VERSION_') &&
+        !_isCommented(line));
+    // Either does not set kotlinOptions or does and uses non-string based syntax.
+    if (hasKotlinOptions && !kotlinOptionsUsesJavaVersion) {
+      // Bad lines contains the first 4 lines including the kotlinOptions section.
+      String badLines = '';
+      final int startIndex =
+          gradleLines.indexOf(gradleLines.firstWhere(isKotlinOptions));
+      for (int i = startIndex;
+          i < math.min(startIndex + 4, gradleLines.length);
+          i++) {
+        badLines += '${gradleLines[i]}\n';
+      }
+      final String kotlinErrorMessage = '''
+If build.gradle(.kts) sets jvmTarget then it must use JavaVersion syntax.
+  Good:
+    android {
+      kotlinOptions {
+          jvmTarget = JavaVersion.VERSION_$_minimumJavaVersion.toString()
+      }
+    }
+  BAD:
+    $badLines
+''';
+      printError(
+          '$indentation${kotlinErrorMessage.split('\n').join('\n$indentation')}');
+      return false;
+    }
+    // No error condition.
+    return true;
+  }
+
+  bool _validateJavaKotlinCompileOptionsAlignment(List<String> gradleLines) {
+    final List<String> javaVersions = <String>[];
+    // Some java versions have the format VERSION_1_8 but we dont need to handle those
+    // because they are below the minimum.
+    final RegExp javaVersionMatcher =
+        RegExp(r'JavaVersion.VERSION_(?<javaVersion>\d+)');
+    for (final String line in gradleLines) {
+      final RegExpMatch? match = javaVersionMatcher.firstMatch(line);
+      if (!_isCommented(line) && match != null) {
+        final String? foundVersion = match.namedGroup('javaVersion');
+        if (foundVersion != null) {
+          javaVersions.add(foundVersion);
+        }
+      }
+    }
+    if (javaVersions.isNotEmpty) {
+      final int version = int.parse(javaVersions.first);
+      if (!javaVersions.every((String element) => element == '$version')) {
+        const String javaVersionAlignmentError = '''
+If build.gradle(.kts) uses JavaVersion.* versions must be the same.
+''';
+        printError(
+            '$indentation${javaVersionAlignmentError.split('\n').join('\n$indentation')}');
+        return false;
+      }
+
+      if (version < _minimumJavaVersion) {
+        final String minimumJavaVersionError = '''
+build.gradle(.kts) uses "JavaVersion.VERSION_$version".
+Which is below the minimum required. Use at least "JavaVersion.VERSION_$_minimumJavaVersion".
+''';
+        printError(
+            '$indentation${minimumJavaVersionError.split('\n').join('\n$indentation')}');
+        return false;
+      }
+    }
+
     return true;
   }